requirements.txt
einops==0.8.0
transformers==4.43.3
huggingface-hub
diffusers==0.30.1
sentencepiece==0.2.0
gradio==5.22.0
torch==2.4.0+cu124
torchvision==0.19.0+cu124
--extra-index-url https://download.pytorch.org/whl/cu124
accelerate
Pillow
